Move module `obj::signal` -> `signal`
Remove mentions of module in use-statements and doc links.
Retain `#[doc(hidden)]` re-export for SemVer compat.
