+++ godot-core/src/private.rs
@@ -30,12 +30,12 @@ mod reexport_pub {
     #[cfg(feature = "trace")]
     pub use crate::meta::{CowArg, FfiArg, trace};
     pub use crate::obj::rtti::ObjectRtti;
-    pub use crate::obj::signal::priv_re_export::*;
     pub use crate::registry::callbacks;
     pub use crate::registry::plugin::{
         ClassPlugin, DynTraitImpl, ErasedDynGd, ErasedRegisterFn, ITraitImpl, InherentImpl,
         PluginItem, Struct,
     };
+    pub use crate::signal::priv_re_export::*;
     pub use crate::storage::{
         IntoVirtualMethodReceiver, RecvGdSelf, RecvMut, RecvRef, Storage, VirtualMethodReceiver,
         as_storage,
