+++ godot-core/src/signal/signal_connections_registry.rs
