+++ godot-core/src/signal/signal_object.rs
