+++ godot-core/src/signal/signal_receiver.rs
@@ -17,8 +17,7 @@ use crate::obj::{Gd, GodotClass};
 
 /// Trait that is implemented for functions that can be connected to signals.
 ///
-/// This is used in [`ConnectBuilder`](crate::obj::signal::connect_builder::ConnectBuilder).
-/// There are three variations of the `C` (class instance) parameter:
+/// This is used in [`ConnectBuilder`]. There are three variations of the `C` (class instance) parameter:
 /// - `()` for global and associated ("static") functions.
 /// - `&C` for `&self` methods.
 /// - `&mut C` for `&mut self` methods.
@@ -29,8 +28,8 @@ use crate::obj::{Gd, GodotClass};
 /// To write generic code that handles different functions and forwards them to [`TypedSignal`] and [`ConnectBuilder`] APIs,
 /// you have to use `SignalReceiver` in combination with [`IndirectSignalReceiver`]. Consult the docs of the latter for elaboration.
 ///
-/// [`TypedSignal`]: crate::obj::signal::TypedSignal
-/// [`ConnectBuilder`]: crate::obj::signal::ConnectBuilder
+/// [`TypedSignal`]: crate::signal::TypedSignal
+/// [`ConnectBuilder`]: crate::signal::ConnectBuilder
 ///
 /// # Hidden `impls` in this doc
 /// To keep this documentation readable, we only document one variant of each `impl SignalReceiver`: arbitrarily the one with three parameters
@@ -47,8 +46,8 @@ pub trait SignalReceiver<C, Ps>: 'static {
 /// If you want to write generic code that can handle different functions and forward them to [`TypedSignal`] and [`ConnectBuilder`] APIs,
 /// you have to use [`SignalReceiver`] in combination with this type. Please check its documentation for a detailed explanation.
 ///
-/// [`TypedSignal`]: crate::obj::signal::TypedSignal
-/// [`ConnectBuilder`]: crate::obj::signal::ConnectBuilder
+/// [`TypedSignal`]: crate::signal::TypedSignal
+/// [`ConnectBuilder`]: crate::signal::ConnectBuilder
 ///
 /// # Background
 ///
